fix(money): re-land the Decimal migration — 210 float-money violations to 0 [AITBC-152]
Squashes PRs #161-#165, which Gitea marked merged but whose content never
reached main. Verified by content rather than SHA: main still had
no_float_money_baseline.json at "total": 210, no cli/aitbc_cli/utils/money.py,
none of the four Alembic revisions, and a v0.23 release log ending at V23-33.
The commits survived only as dangling objects; they are tagged recovered/* .

Rebased onto current main (through PR #179) and applied with no conflicts.
V23-34 through V23-40 in docs/releases/v0.23/release.log.

  coordinator-api  168 -> 0   c7d1f4a9e230 converts 102 columns across 45 tables
  blockchain-node   24 -> 0   d4e8b91c0a37 converts two GPU resource columns
  cli               11 -> 0
  the rest          13 -> 0   f2b6c04a91d8, a3e7c15b8d94
  parameters       131 -> 0

Six widenings of scripts/lint/no_float_money.py, each of which found real
defects in code the previous pass had reported clean: annotations; `per` and
`target` in DERIVED_TOKENS hiding price_per_hour and target_amount; camelCase;
ast.IfExp; ast.Attribute annotation targets hiding self.earnings and
self.total_spent; and parameter annotations. "Zero violations" only ever means
zero of what the checker can currently see, so the guard now says so in its
docstring and the baseline records that it started at 210.

The largest single defect was in ait_to_seconds, which turns a user's --amount
into the integer compute-seconds the chain settles:

    int(0.5025 * 3600) == 1808        0.5025 * 3600 == 1808.9999999999998
    int(Decimal("0.5025") * 3600) == 1809

1402 of the million four-decimal inputs between 0.0001 and 100.0000 lost a
compute-second that way, always the sender's. tests/unit/test_ait_unit_conversion.py
pins both the fix and the old behaviour.

@click.option(type=float) made CLI annotations unenforceable, since Click
converts before the handler runs, so cli/aitbc_cli/utils/money.py adds a
DecimalParamType used by every money option. Local wallet, multisig and
exchange JSON now store money as decimal strings, read back through
wallet_amount(), which also accepts the numbers older builds wrote.

Money that crosses a signature or hash boundary deliberately stays float, each
site carrying a `# not-money:` marker with its reason -- Decimal is not
JSON-serialisable and "0.5" != 0.5 changes the hash, so converting those is a
hard fork. Same for quantities whose names carry a money token but which are
rates, ranking weights, timeouts, RL returns or numpy statistics.

ruff, mypy aitbc/, the 12-app mypy ratchet and the guard all pass. 2672 passed
across tests/unit+cli+services+security; blockchain-node 699, trading 84, miner
60, agent-coordinator 76, pool-hub 41, agent-sdk 25, gpu 17, edge 11, crypto 2.
coordinator-api's 48 failures are byte-identical to plain origin/main, verified
in a detached worktree.

+++ apps/coordinator-api/src/coordinator_api/agent_identity/sdk/models.py
@@ -5,6 +5,7 @@ Data models for the Agent Identity SDK
 
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 from enum import StrEnum
 from typing import Any
 
@@ -97,9 +98,9 @@ class AgentWallet:
     chain_address: str
     wallet_type: str
     contract_address: str | None
-    balance: float
-    spending_limit: float
-    total_spent: float
+    balance: Decimal
+    spending_limit: Decimal
+    total_spent: Decimal
     is_active: bool
     permissions: list[str]
     requires_multisig: bool
@@ -211,7 +212,7 @@ class TransactionRequest:
     """Request model for transaction execution"""
 
     to_address: str
-    amount: float
+    amount: Decimal
     data: dict[str, Any] | None = None
     gas_limit: int | None = None
     gas_price: str | None = None
@@ -246,10 +247,10 @@ class WalletStatistics:
 
     total_wallets: int
     active_wallets: int
-    total_balance: float
-    total_spent: float
+    total_balance: Decimal
+    total_spent: Decimal
     total_transactions: int
-    average_balance_per_wallet: float
+    average_balance_per_wallet: Decimal
     chain_breakdown: dict[str, dict[str, Any]]
     supported_chains: list[str]
